I moved the mapper package to the utils package
